Fixing the timer bug, and adding minor variation to load

+++ Cell.h
@@ -41,7 +41,7 @@ class Cell : public CBase_Cell {
     int stepCount;		// to count the number of steps, and decide when to stop
     int myNumParts;   //number of atoms in my cell
     int inbrs;        //number of interacting neighbors
-    int stepTime;
+    double stepTime;
     int updateCount;
     double energy[2]; //store kinetic energy - initial and final
 	int numReadyCheckpoint;
